added separate files for customer classes
