+++ dotnet-testing/FluentAssertions/FluentAssertionsTests/SpecialCustomer.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FluentAssertionsTests
+{
+    public class SpecialCustomer : Customer
+    {
+        public SpecialCustomer() { }
+    }
+}
